+++ Engine/Source/Editor/UILayers/Inspector.h
@@ -1,5 +1,6 @@
 #include "ImGui/ImGuiBaseLayer.h"
-#include "UILayers/ImGuiUtils.hpp"
+#include "ImGui/ImGuiUtils.hpp"
+
 namespace editor
 {
